Split up rest of pyflakes.rules
